LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerlock/pointerevent_pointerlock_supercedes_capture-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Events pointer lock test

Follow the test instructions with mouse. If you don't have the device skip it.

Test Description: This test checks that we do not set the pointer capture when any element in the page gets a pointer lock.
Press left button down on the green rectangle and hold it.
Move the mouse inside the green rectangle.
Test passes if the pointer capture is not set when the green rectangle gets the pointer lock.


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN mouse no pointercapture while pointerlock
